requirement.txt
certifi==2018.1.18
chardet==3.0.4
dj-pagination==2.3.2
Django==1.11
djangorestframework==3.7.7
get==0.0.39
idna==2.6
post==0.0.26
public==0.0.65
pytz==2017.3
query-string==0.0.28
request==0.0.26
requests==2.18.4
urllib3==1.22

Flask==0.10.1
Flask-SocketIO
Jinja2==2.7.2
MarkupSafe==0.18
Werkzeug==0.9.4
gevent==1.0
gevent-socketio==0.3.6
gevent-websocket==0.9.2
greenlet==0.4.2
itsdangerous==0.23
ujson==1.33
